fix(helm): configure version for snapshot api (#361)
On some managed K8s clusters like OpenShift or GKE that ship with
their own default CSI Drivers, the controllers will end up creating
v1beta1 objects. An issue like the following was noticed when running
cstor charts on GKE 1.19.

```
 unable to recognize "": no matches for kind "VolumeSnapshotClass" in
 version "snapshot.storage.k8s.io/v1"
```

This commit uses the helm .Capabilities to check the version available
in the cluster and install the class accordingly. As the .Capabilities
might return false when using with template command, v1 is kept for false
(or default) case.

Signed-off-by: kmova <kiran.mova@mayadata.io>